[CHORE] MemberService 분리
