+++ storage/db-core/src/main/java/com/org/gunbbang/repository/MemberBreadTypeRepository.java
@@ -3,6 +3,8 @@ package com.org.gunbbang.repository;
 import com.org.gunbbang.entity.Member;
 import com.org.gunbbang.entity.MemberBreadType;
 import java.util.List;
+
+import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -11,6 +13,7 @@ public interface MemberBreadTypeRepository extends JpaRepository<MemberBreadType
   List<MemberBreadType> findAllByMember(Member member);
 
   @Query("SELECT mbt FROM MemberBreadType mbt where mbt.member.memberId = :memberId")
+  @Transactional
   List<MemberBreadType> findAllByMemberId(Long memberId);
 
   boolean existsByMember(Member member);
@@ -19,5 +22,6 @@ public interface MemberBreadTypeRepository extends JpaRepository<MemberBreadType
 
   @Modifying
   @Query("delete from MemberBreadType mbt where mbt.member = :member")
+  @Transactional
   void deleteAllByMember(Member member);
 }
