+++ storage/db-core/src/main/java/com/org/gunbbang/repository/MemberNutrientTypeRepository.java
@@ -3,6 +3,8 @@ package com.org.gunbbang.repository;
 import com.org.gunbbang.entity.Member;
 import com.org.gunbbang.entity.MemberNutrientType;
 import java.util.List;
+
+import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -18,6 +20,7 @@ public interface MemberNutrientTypeRepository extends JpaRepository<MemberNutrie
 
   @Modifying
   @Query("delete from MemberNutrientType mnt where mnt.member=:member")
+  @Transactional
   void deleteAllByMember(Member member);
 
   @Query(
